add verif if id exist or is valid to /:id

+++ src/controllers/certificationController.js
@@ -54,6 +54,16 @@ exports.getAllCertifications = async (req, res) => {
 // Modifie une certification par son ID
 exports.updateCertification = async (req, res) => {
   try {
+    // Recherche si l'id est vide
+    if (req.params.id == undefined || req.params.id.trim() == "") {
+      return res.status(400).json({ error: 'Empty' });
+    }
+    // Recherche si l'id est correcte
+    const skillIdRegex = /^[0-9a-fA-F]{24}$/;
+    if (!skillIdRegex.test(req.params.id.trim())) {
+      return res.status(400).json({ error: 'Not an ID' });
+
+    }
     // Rechercher la certification à mettre à jour
     const certification = await Certification.findById(req.params.id);
     if (!certification) {
@@ -95,14 +105,24 @@ exports.updateCertification = async (req, res) => {
 // Récupère une certification par son ID
 exports.getCertificationById = async (req, res) => {
   try {
-    // Rechercher la certification dans la base de données par son ID
-    const certification = await Certification.findById(req.params.id);
+    // Recherche si l'id est vide
+    if (req.params.id == undefined || req.params.id.trim() == "") {
+      return res.status(400).json({ error: 'Empty' });
+    }
+    // Recherche si l'id est correcte
+    const skillIdRegex = /^[0-9a-fA-F]{24}$/;
+    if (!skillIdRegex.test(req.params.id.trim())) {
+      return res.status(400).json({ error: 'Not an ID' });
+
+    }
 
     // Vérifier si la certification existe
     if (!certification) {
       // Si la certification n'est pas trouvée, renvoyer une réponse avec le code 404
       return res.status(404).json({ error: 'Certification not found' });
     }
+    // Rechercher la certification dans la base de données par son ID
+    const certification = await Certification.findById(req.params.id);
 
     // Si la certification est trouvée, renvoyer une réponse avec la certification
     res.status(200).json(certification);
@@ -117,14 +137,24 @@ exports.getCertificationById = async (req, res) => {
 // Supprime une certification par son ID
 exports.deleteCertification = async (req, res) => {
   try {
-    // Rechercher la certification à supprimer dans la base de données par son ID et la supprimer
-    const certification = await Certification.findByIdAndDelete(req.params.id);
+    // Recherche si l'id est vide
+    if (req.params.id == undefined || req.params.id.trim() == "") {
+      return res.status(400).json({ error: 'Empty' });
+    }
+    // Recherche si l'id est correcte
+    const skillIdRegex = /^[0-9a-fA-F]{24}$/;
+    if (!skillIdRegex.test(req.params.id.trim())) {
+      return res.status(400).json({ error: 'Not an ID' });
+
+    }
 
     // Vérifier si la certification existe
     if (!certification) {
       // Si la certification n'est pas trouvée, renvoyer une réponse avec le code 404
       return res.status(404).json({ error: 'Certification not found' });
     }
+    // Rechercher la certification à supprimer dans la base de données par son ID et la supprimer
+    const certification = await Certification.findByIdAndDelete(req.params.id);
 
     // Si la certification est trouvée et supprimée avec succès, renvoyer une réponse avec le code 200
     res.status(200).send('Certification deleted');
